+++ src/shot.h
@@ -39,7 +39,7 @@ namespace GameEntities {
 
     class Shot : public GameEntity {
     public:
-        Shot(double x, double y, double dx, double dy, bool active, Game::Game* game)
+        Shot(int x, int y, int dx, int dy, bool active, Game::Game* game)
             : GameEntity(x, y, dx, dy, active, game)
         {
             image = game->get_image("shot.png");
